JRUBY-143: RubyObject.hashCode fails for hash values not in int range
git-svn-id: http://svn.codehaus.org/jruby/trunk/jruby@2243 961051c9-f516-0410-bf72-c9f7e237a7b7

+++ src/org/jruby/RubyObject.java
@@ -809,7 +809,7 @@ public class RubyObject implements Cloneable, IRubyObject {
     }
     
     public final int hashCode() {
-    	return RubyNumeric.fix2int(callMethod("hash"));
+    	return (int) RubyNumeric.fix2int(callMethod("hash"));
     }
 
     /** rb_obj_type
